Make clauses in internal package abstract
