+++ thingif/src/main/java/com/kii/thingif/internal/clause/Range.java
@@ -7,7 +7,7 @@ import android.support.annotation.NonNull;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class Range implements Clause {
+public abstract class Range implements Clause {
 
     protected final String field;
     protected final Long upperLimit;
@@ -25,58 +25,6 @@ public class Range implements Clause {
         this.upperIncluded = upperIncluded;
         this.lowerLimit = lowerLimit;
         this.lowerIncluded = lowerIncluded;
-    }
-    public static  Range range(
-            @NonNull String field,
-            Number upperLimit,
-            Boolean upperIncluded,
-            Number lowerLimit,
-            Boolean lowerIncluded) {
-        return new Range(
-                field, upperLimit.longValue(),
-                upperIncluded,
-                lowerLimit.longValue(),
-                lowerIncluded);
-    }
-    public static Range greaterThan(
-            @NonNull String field,
-            @NonNull Number lowerLimit) {
-        return new Range(
-                field,
-                null,
-                null,
-                lowerLimit.longValue(),
-                Boolean.FALSE);
-    }
-    public static Range greaterThanEquals(
-            @NonNull String field,
-            @NonNull Number lowerLimit) {
-        return new Range(
-                field,
-                null,
-                null,
-                lowerLimit.longValue(),
-                Boolean.TRUE);
-    }
-    public static Range lessThan(
-            @NonNull String field,
-            @NonNull Number upperLimit) {
-        return new Range(
-                field,
-                upperLimit.longValue(),
-                Boolean.FALSE,
-                null,
-                null);
-    }
-    public static Range lessThanEquals(
-            @NonNull String field,
-            @NonNull Number upperLimit) {
-        return new Range(
-                field,
-                upperLimit.longValue(),
-                Boolean.TRUE,
-                null,
-                null);
     }
 
     public String getField() {
@@ -155,17 +103,7 @@ public class Range implements Clause {
         this.lowerLimit = (Long)in.readValue(Range.class.getClassLoader());
         this.lowerIncluded = (Boolean)in.readValue(Range.class.getClassLoader());
     }
-    public static final Creator<Range> CREATOR = new Creator<Range>() {
-        @Override
-        public Range createFromParcel(Parcel in) {
-            return new Range(in);
-        }
 
-        @Override
-        public Range[] newArray(int size) {
-            return new Range[size];
-        }
-    };
     @Override
     public int describeContents() {
         return 0;
