+++ thingif/src/main/java/com/kii/thingif/trigger/clause/Clause.java
@@ -1,8 +1,4 @@
 package com.kii.thingif.trigger.clause;
 
-import android.os.Parcelable;
-
-import org.json.JSONObject;
-
 public interface Clause extends com.kii.thingif.internal.clause.Clause {
 }
